hooks: removed Linux Kernel LightNVM hooks (lnvm + pblk)
This goes in concert with adding the removed hooks to the
`cijoe-pkg-lightnvm`.

The motivation is that the base of CIJOE should be a simple toolchain
providing a test-runner and reporter of Bash and Python testcases.
Different use-cases are put into packages, such as having things
concerned with the Linux Lightnvm subsytem / PBLK in its own package,
e.g. the `cijoe-pkg-lightnvm`.

Signed-off-by: Simon A. F. Lund <os@safl.dk>