replace hook-locking with runner-managed file-lock
It is a runner task to ensure that the runner is not executing when it
is already running. Putting it into a hook was somewhat convenient but
was really hard for users to understand how to get rid of a lock after a
failure. Also, for actually communicating that CIJOE was locked was not
that easy to grasp since it had started executing.

With the runner-managed lock, nothing runs until the runner itself
manages to take the lock and it produces better error-messages /
guidance for the user.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>