mod:py:runner: fix mypy errors
Note that the error handling flow is also changed: errors during
initialization of the TestRun are now raised as exceptions instead of
returning None or 1.

This is done in order to avoid requiring setting optional types for all
of the setup functions, but has the very nice side effect of making sure
that errors are caught and communicated to the end user in a single
place, namely the try-except in `main`.

Signed-off-by: Michael Bang <mi.bang@samsung.com>