mod:py:runner: add counter to TestRun
As there can be multiple instances of testcases, testsuites, and
testplans, then a way to distinguish them from each other is needed. And
preferably in a way which provides ordering when lexicographically
sorted.
Currently for testuites, this is done by forwarding an enumeration value
during testsuite-setup.
The intent of this TestRun.count and TestRun.inc() are to provide a
global counter for of these instead of parsing around enumeration
values.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>