Add support for multiple testplans in a single testrun
The CLI provided by ``cij_runner`` now takes an extendable list of
testplans. Additionally, the ``testplan`` and ``env`` are now named
arguments, provided by ``--testplan`` and ``--env``.

A TestPlan abstraction is added to the runner, it is a extraction of the
testplan properties of the TestRun class. The testrun now has a list of
TestPlans instead of a list of TestSuites.

TestSuite setup now utilizes the TestPlan.res_root instead of the
TestRun.res_root. Thus, output is nested further down, from the root of
testplans.

The reporter is extended with post-processing of the TestPlans and the
report-template is similarly extended with iterating over testplans.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>