mod:py: add configuration class
The different tools require that CIJOE environment variables are
set; CIJ_ROOT, CIJ_TESTCASES, CIJ_TEMPLATES, etc. They also each do
their own parsing to setup a conf-dict.

This introduces a Config-class to hold these paths and a
factory-function to contruct it. The factory function inspects the
environment variables, and when they are not set, constructs the paths
based on the root provided by ``cij_root``.

This will allow for tools to run without having to start the interactive
cijoe-mode first.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>