mod:py:runner: persist state at end of suite and plan
The status of TestSuite is derived after all testcases in the suite have
finished and the status of TestPlan derived after all testuites within
the plan have completed. However, the testrun is persisted after each
testcase is completed, thus one has to wait for a testcase in the *next*
testsuite or testplan to complete in order to get the status.

This is a bit annoying when doing extraction, analyses or just reporting
while the runner is doing its thing. So, thus the state i persistent
more frequently.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>