bin:plotter: add tool for plotting metrics
Basic implementation to determine what might be additional features or
requirements to extraction / normalization phase. Here is what the
plotter does:

1. Collects metrics from a testrun
2. Transform the metrics to data for a specific plot e.g. line-plot
3. Render a specific plot (line-plot)

Things to consider:

* type-conversion of e.g. fio bs and iodepth as integers for ordering
e.g. as a function of iodepth, possibly also a need-to do
unit-normalization for things such as bs, e.g. the 'k' in fio bs is
binary actually KiB, not short-hand for thousand.

And for the plotter itself:

* Specification of context-portion of query for x-axis
* Specification of context-portion of query for reducing dataset
* How to load plotters in other cijoe packages and provided directly

Making a generic plotting-tool is infeasible, as even in this plotter,
it quickly becomes a command-line interface to a plotting API. It would
seem that the plotter should just have some basic convenient plots e.g.
a line-plot and a grouped bar-plot.
Then if the user wants some else, then it should be possible to provide
user-defined implementations of steps 2. and 3.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>