Add a naive path sanitizer
When uploading artifacts via GitHUB actions, then the output-directories
generated by cijoe are not allowed. Also, some parametrize generate
paths like '/dev/nvme0n1', without sanitization, then this makes cijoe
create a strange directory structure.

To fix the above, a naive sanitize_ident() function is added and
utilized by cijoe.set_output_ident(). This is not a thorough solution,
but it is sufficient to fix the above mentioned issues.  One downside is
that, with this, then the directory-structure of folder-organized
pytests are not keeping that structure in the workflow-output.

Signed-off-by: Simon A. F. Lund <os@safl.dk>