cli: change default behavior
After using the cli for a while, it is clear that the common-case is
invocation as in:

  cijoe -fp ..

That is, move the current 'cijoe-outout' into 'cijoe-archive' and
produce a report after the workflow is done. So making this behavior the
default.

* This also avoids the unexpected "aborting..." log-message when running
  it with 'cijoe -p'.
* The intent of preventing the user from re-running while currently
  executing has never worked as intended, as users just always add the
  '-f' or just remove the output directory before running.

Thus, the default behavior is changed, the '-f' is removed and in case
somebody does not want a report at the end, then a new option is added
'-n'/'--no-report'.

Signed-off-by: Simon A. F. Lund <os@safl.dk>