cli,core,transport: reduce verbosity of logging
The logging-info produces by paramiko was spamming with traceback info.
This changes it to only output at CRITICAL. Additionally, the exceptions
are captured closer to the source and logged as a one-liner to with
level(error), and traceback with level(DEBUG).

This is done to not be "overwhelmed" with logging information on a
regular invocation of cijoe.

Signed-off-by: Simon A. F. Lund <os@safl.dk>