worklet/reporter: adjust report-opening for macOS
To open a workflow-report, then the 'reporter-worklet' uses the Python
'webbrowser' module. This is done, with the intent that it would provide
a cross-platform capabable means of opening the preferred webbrowser.

Unfortunately then the webbrowser module does not provide a
cross-platform method for opening local files. However, prefixing
'file://' seems to enable the intended behavior on macOS.

Thus, applying it here.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>