cli: add a tag option for a workflow run
It would be useful to give a tag for a workflow run
while running cijoe during development. This option enables
to give a tag that will be stored in the workflow.state. The same
text will be used as a prefix with timestamp while storing the results
in the archive folder.

Example usage:
$ cijoe -c default.config -w example.workflow -t "directIO" --tag "io_uring"
$ ls -la cijoe-archive/
directIO-io_uring-2023-02-07_06:44:34

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>