Encapsulate installation in pipx-venv
With the arrival of PEP 668 on various distributions, then ``pip install
--user`` and general system-wide installation is no longer not just not
recommended, it is also, impractical.
Use of Python via a virtual-environment is more or less required now,
atleast for packages not provided via the system package-manager.

Thus, this change, updating the documentation, describing the
recommended approach of installation cijoe and cijoe-packages via pipx.

And importantly, changing the invocation of pytest, to utilize
pytest-cli instead of module-invocation, as it will otherwise not have
access to the pipx-venv containing both pytest and cijoe.

Signed-off-by: Simon A. F. Lund <os@safl.dk>