Change .config to .toml and .workflow to .yaml
In the case of the workflows, then it is mere change of file-extension /
suffix. For the configuration-files then the format is changed from yaml
to toml.

So why change? Well, yaml is the goto-representation for
workflow-representation in related tools/systems such as CI-definitions
(GitLAB CI and GitHUB Actions) and ansible. Thus, a user with any prior
experience any of these, will have a minimal learning curve when picking
up cijoe.

For configuration, then yaml is usable, however, toml is the goto
representation, e.g the Python pyproject.toml and similar have opted for
toml for configuration-files, due to the "oldchool" and easily readable
ini-file-style, and the deterministic non-cycling representation.

The "custom" filenames ".workflow" and ".config" is a both a niceaty,
however, the ".config" name collides with the common use-case of Linux
kernel configs, thus, it is impractical for the file-collection.

Thus, switching formats to better match the the data-representation
purpose.

Signed-off-by: Simon A. F. Lund <os@safl.dk>