feat(cli): add entry-point named 'joe'
This adds a an entry-point for the command-line tool, named 'joe'. This
is in addition to the existing entry-point named 'cijoe'.

Additionally, then the usage-text is changed to use the name from
sys.argv, this is to correctly show it, as 'cijoe' is now only one of
two possible names of the cli.

Signed-off-by: Simon A. F. Lund <os@safl.dk>