feat(core/scripts): add 'run_local' option to testrunner
The testrunner only did run_local when executing pytest. The "remote"
execution was thus does per usual by cijoe.run() executions in the
pytests. However, there are use-cases where the ability to execute
pytest remote.

This change enables it.

This extends requirements to include 'tomli-w'. This is needed to
created a modified copy of the cijoe-config for run_local==False.

Signed-off-by: Simon A. F. Lund <os@safl.dk>