refactor: introduce naming convention for default config and workflow
The cijoe cli-tool searched the current working directory for config
files (.toml) and workflow files (.yaml) and used the first ones it
found.

The issue with this is that the tool would pick up files that might not
be cijoe configs/workflows, such as Rust / Python project definitions.

Thus, to work around this, a naming convention for "default" configs and
workflows are introduced:

* config: "cijoe-config.toml"
* workflow: "cijoe-workflow.yaml"

Thus, the search for any .toml and any .yaml is replaced with only using
those name according to the default naming. Additionally, then the
default provided with cijoe, and by cijoe packages are changed as well,
such that: ``cijoe --example <pkgname>``  will work accordingly.

Signed-off-by: Simon A. F. Lund <os@safl.dk>