feat(monitor): add dump of command-output
Previously, the monitor would just write the path to command-output
files as their creation were observed. This expands on the functionality
and dumps the content of the output-files as well.

Additionally, then, when no output-files have been dumped in a
WorkflowMonitor.notify_seconds, then a message is displayed that no have
been received.

Worth noting is that the monitor is now a "workflow-utility" instead of
a "workflow-option", and thus, intended to be executed "stand-alone".
That is, another process should be started executing:

	cijoe -m

And not "just" piggy-backed upon a workflow-run.

Signed-off-by: Simon A. F. Lund <os@safl.dk>