fix(monitor): change monitor-output from print() to log()
The intent of the monitor is to provide output from cijoe.run(), that
is, command output, as though one were executing the commands manually.

The added meta-output in the form of "# cijoe.monitor: ..."
print-statements got a bit in the way. Thus, this change, adding them as
log-statements instead, and thus a way for the user to disable the
output.

Signed-off-by: Simon A. F. Lund <os@safl.dk>