feat(cli): expand search for config + workflow
A project using cijoe might provide some cijoe-configs which are usable
in ci environments such as GitHUB, GitLAB, Travis, etc. or for custom
local setups using qemu. However, cijoe-configs for systems with actual
HW, such as local development machines / boxes. For these, configs
within the project repository is not very useful.

In such cases, one would rather use the projects cijoe-configs as a
reference and adjusting a copy for the hardware / system available.

It would be convenient to do:

  cp myproject/cijoe/cijoe-config.toml ~/.config/cijoe/cijoe-config.toml

And then adjust ~/.config/cijoe/cijoe-config.toml to describe the
hardware / system. However, cijoe does not by default search for files
in these locations. To enable this, then this change expands the
config-file and workflow-file search to look in:

* cwd
* cwd / .cijoe
* $HOME / .cijoe
* $HOME / .config / cijoe

Thus providing config-search in tree as well as out-of-tree.

Signed-off-by: Simon A. F. Lund <os@safl.dk>