mk: Add --force when installing
make install may not be able to succeed if cijoe is previously
installed. Add --force to the pipx arguments to make sure
the package reinstalls in this case.

Fixes warning:
'cijoe' already seems to be installed. Not modifying existing
installation in '/Users/dagomez/.local/pipx/venvs/cijoe'. Pass
'--force' to force installation.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>