feat(cli): allow execution of cijoe-script with cijoe
cijoe-scripts need to be executed with the Python interpreter that has the
virtual environment loaded in which cijoe is installed. If the user does not
know the path to this Python interpreter, cijoe-scripts can now be run with
cijoe, i.e. `cijoe cijoe-script.py`. Multiple cijoe-scripts can be added as
argument, to run them as a single workflow, e.g.
`cijoe script1.py script2.py script3.py`.

Solves second part of issue #73.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>