fix(gha,mk): ensure test of different python versions
The installation of cijoe in gha reliad of the 'make all' target, it in
turn invoked 'make instal' which calls 'pipx install'. However, it was
not invoked with '--python', e.g. 'pipx install ... --python python3'.
Without this, then it will use the system default, instead of the python
version installed in the github action.
Thus, to actually, and properly, test the different Python versions,
then this change is needed.

Signed-off-by: Simon A. F. Lund <os@safl.dk>