docs: re-organize documentation
More topics are put "up front", that is, the most important ones
like scripts and workflows. Also, instead of separating "usage" and
"definition" of these, then they are covered in self-contained sections.

Then, the most essential stuff is easy to find, and the less essential
stuff e.g. perf. reqs. are in subsections.

Also, combined the "landing-page" and the "introduction" since it seemed
a bit redundant.

Signed-off-by: Simon A. F. Lund <os@safl.dk>