feat(transport): allow defining multiple transport endpoints
Added the `endpoint_name` parameter to `cijoe.run(...)`, which allows users to
specify which endpoint they want to execute commands on. Endpoints are defined
in the configuration file under the key `cijoe.transport` with user-defined
names. If the `endpoint_name` parameter is not given, the command is executed
on the first defined endpoint. If no endpoints are defined, the command is
executed on the initiator.

Solves issue #75

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>