feat(qemu): add checksum checks on qemu images
Before, images were not downloaded if a file on the given path already
existed. Now, checksums are compared to determine whether
the image should be downloaded. Configuration files must be expanded with
a qemu.guest.<init>.url_checksum which points to the checksum file of the
downloaded image.

Solves issue #83

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>