feat(gha): add verification of the cijoe package, fio
The example configuration file is split into build and benchmark
sections to simplify examples, ensuring each file contains only entries
relevant to the specific use case.

Additional work is required on the package. This change exercises the
build, install, and check scripts, but the more complex fio-wrapper
script is not yet covered and the documentation needs updating as well.

Signed-off-by: Simon A. F. Lund <os@safl.dk>