refactor(gha,cli): improve example discovery and organization
The previous method of discovering examples was cumbersome. Users had
to run `cijoe --resources` and manually search for files with "example"
in their names, then piece together configuration files, workflows,
and scripts.

This refactor changes the behavior of the `cijoe --example` command to
make example discovery more intuitive:

- Running `cijoe --example` now lists all available examples.
- Running `cijoe --example qemu` outputs all examples specific to the `qemu` package.
- Running `cijoe --example qemu.build` outputs only the `qemu.build` example.

Additionally, examples are now output into a subfolder of the current
working directory in the format:

  cijoe-example-<pkg_name>.<example>

This improves usability and streamlines access to examples.

The GHA workflow for packages is adjusted to the above changes. Also,
the format of the 'usage_examples' is changed to a multi-line-list.

Signed-off-by: Simon A. F. Lund <os@safl.dk>