refactor(gha,pkg/qemu): replace 'default' with two explicit examples
The 'default' example demonstrated how to initialize and start a guest
using cloud-init, but previously included a commented-out example of
using bootimage/diskimage instead.

To enhance usability and simplify testing, the example has been
refactored into two separate examples, with the commented-out portion
removed. This change improves clarity and makes the examples easier to
use and test.

Furthermore, both examples have been added to the GHA workflow to ensure
they work as expected. Previously, only the 'build' example was included
in the workflow.

Signed-off-by: Simon A. F. Lund <os@safl.dk>