feat(cli): add subparsers for workflows and scripts
This commit changes the way cijoe scripts and workflows are run. The
interface is now `cijoe [target] [...options]`, where `target` is either a path
to a stand-alone cijoe script, the identifier for a cijoe-script, or the path to
a cijoe workflow.

If `target` is referencing a stand-alone script, additional cli args can
be defined by adding a function to the cijoe script file of the form
`add_subparser_args(parser: ArgumentParser)`.
This will allow us to remove the `step` argument in the cijoe script main function.

If `target`is referencing a workflow file, additional positional arguments will be
treated as step arguments, defining which steps in the workflow will be run.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>