fix(core/workflow): add linting for workflow step.with
In order to pass the step.with keys to the args Namespace, we need
to ensure that these do not overwrite the attributes set by the ArgumentParser.
Therefore we disallow cli options as step.with keys in workflow.

Furthermore, we ensure that the type of the elements in the step.with
dict can only be primitive types (numbers. strings, booleans) or list
of primitive types.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>