fix(core/cli): use parse_args to merge existing args with step args
Running cijoe scripts from the command line with cli-arguments failed because
`parse_args()` were called twice (once when evaluating the command, and once when
executing the step in the ad-hoc script), and therefore created duplicate args in
the Namespace. By using `parse_args()` with the namespace parameter,
we merge the existing namespace with the script args more fluently, only setting
the default values for args that were not given in the cli.

We keep the `script_args` variable to allow us to remove script arguments
after execution of the script. This way, if a later step has not set a specific
argument, the default value will be correctly set.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>