fix(core/scripts): change bool cli args to "true"/"false" choices
In the argparser documentation, it is mentioned that the `bool()` function is not
recommended as type converter, as it only evaluates the empty string to False,
and everything else to True - meaning that "false" would be evaluated to
True.

By using the argparser `choices` and a custom StoreAction, we allow the
user to implement boolean arguments with default=True. Furthermore,
aguments like `--run_local false` will correctly be evaluated to False.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>