fix(core/testrunner): remove whole transport item in config if `run_local=false`
With the addition of multiple transports, a transport not named "ssh" will not be
removed from the config file when `run_local=false` as intended. With this commit
the whole transport item is removed, ensuring that cijoe does not try to re-connect
to another transport when running tests on a remote host.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>