fix(cli/loglevel): add force param to `log.basicConfig()`
According to the docs, the `log.basicConfig()` method does nothing if the root
logger already has handlers configured. Apparently this happens before the
method is called in the code, and therefore we need the `force` parameter to
ensure the configuration is applied.
https://docs.python.org/3/library/logging.html#logging.basicConfig

Additionally, the call to `basicConfig()`
is moved to just after the `parse_args()` call to ensure the configuration is
applied for the whole run.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>