feat(system_imaging): add cloud-init example for fbsd
nuage-init is the initialisation tool used in FreeBSD images (they call
it "basic cloudinit"). This is a subset of cloud-init, and doesn't
support all the same features. To support FreeBSD images, we host these
images with cloud-init pre-installed and provide the download URL for
these in the example config files.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>