fix(cli/config): merge configs when multiple are given
If multiple configs were given and more than one defined a subkey of the
same dict, the whole key would be overwritten. Example:

[foo]
bar = 1
baz = 2

[foo]
bar = 3

This would overwrite the whole `foo` dictionary, removing the `baz` key
which is not the intuitive.

With this fix, the toml files are recursively merged so only specified
keys in subsequent config files are overwritten (or added if not defined
in the first config).

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>