refactor(system_imaging): simplify examples and move test of all images
The two examples in the system_imaging package were constructed to test
all images in the given configs. These examples become comprehensive as
we add more and more examples. The purpose was to ensure that all images
defined in the config are valid.

To not create examples that are so large that they break the CI, we
simplify the system_imaging example to just match on the pattern
"*debian*", keeping them 'light-weight'. The test of all images in the
config is moved to a separate GHA job.

This also removes the duplicate example config files in the
system_imaging package, so we don't have to maintain to identical files.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>